refactor: ChessBoardDto에 변환 로직 제거하고 BoardToString.convert()로 변환하도록 변경
